Remove Tune price
Upfront/standardized pricing was causing me more headaches than it was worth. I originally thought it would help me speed up negotiation/spend less time on selling, but it's actually only done the opposite.

It sounds "sales-y" to say but putting the fee this up-front really does cause people to focus more on the price than on the value I'm providing. I can lay out how I'm going to save somebody 60,000 in the next year and also improve their customer's experiences, but still get a reaction like "that's outside our budget". It's putting the cart before the horse. There's a good reason why most high-value services don't put their prices upfront and, while I tried to resist, I have now learned why.

This also gives me more freedom to expand and contract the scope of a Tune as I see fit and to react to the application brought to me. It's easier to tune a 2 year old app than it is an 8 year old one with 100k users, that's just facts. And of course, I'll be delivering even more value to the latter than the former.