Maintain -> help maintain Puma
I've decide that the ambiguity of "maintain <PROJECT>" is yuck if you're not the sole contributor to that project. Saying "I maintain X" creates an ambiguity as to whether or not you're the only maintainer that I'm no longer comfortable with. The truth is that I was never the sole maintainer of Puma and I want to create as many maintainers of Puma as I can.